requirements.txt
streamlit
streamlit-authenticator==0.2.3
mysql-connector-python==8.3.0
python-dotenv==1.0.1
passlib==1.7.4
pandas
Pillow
numpy
PyYAML
geopy
folium
streamlit-folium
requests
streamlit-option-menu
streamlit-geolocation